insufficient testing of dash while character moves
After checking with more enemy objects, I noticed that since the target
position in the enemy class is based on the player position, it
continues to reference that position the player moves. I forgot that I
wanted to take a "snapshot" of the player's screen pos, at the
initiation of the dash.