Pending confirmation, but dash bug with movement appears fixed
I was hoping to avoid this kind of fix, but what I found I needed to do was update the targetPos variable in one of the enemies tick states based on how the player character drove movement. The only issue with that was the fact that the BaseCharacter::Tick() function zeroes its velocity. So I created a private Vector2 called lastVelocity to store the player character's velocity value before it gets zeroed. Through some additional tweaking in the enemy dash function by normalizing the movement imposed upon targetPos by the player character, I think it's starting to feel really good.